requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1
django-bootstrap4==24.4
django-cors-headers==4.4.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
Faker==30.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
ndg-httpsclient==0.5.1
pillow==10.4.0
pusher==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zope.interface==7.1.1
